+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added support for Chart.js 4
 - Fixed tooltip for negative values with `bytes`
 - Dropped support for AJAX through jQuery and Zepto
+- Dropped support for Chart.js 3
 
 ## 4.2.0 (2022-05-25)